+++ src/main/java/io/luchta/forma4j/reader/xml/XmlReader.java
@@ -1,9 +1,9 @@
-package io.luchta.forma4j.reader.compile;
+package io.luchta.forma4j.reader.xml;
 
 import io.luchta.forma4j.context.syntax.SyntaxError;
 import io.luchta.forma4j.context.syntax.SyntaxErrors;
-import io.luchta.forma4j.reader.compile.parse.Parser;
-import io.luchta.forma4j.reader.compile.relation.RelationAnalysis;
+import io.luchta.forma4j.reader.xml.parse.Parser;
+import io.luchta.forma4j.reader.xml.relation.RelationAnalysis;
 import io.luchta.forma4j.reader.model.tag.TagTree;
 import org.xml.sax.SAXException;
 
@@ -11,7 +11,7 @@ import javax.xml.parsers.ParserConfigurationException;
 import java.io.IOException;
 import java.io.InputStream;
 
-public class FormaReaderCompiler {
+public class XmlReader {
 
     public TagTree compile(InputStream xml, SyntaxErrors syntaxErrors) throws ParserConfigurationException, IOException, SAXException {
