+++ src/main/java/io/luchta/forma4j/reader/xml/parse/tagbuilder/TagBuilder.java
@@ -1,4 +1,4 @@
-package io.luchta.forma4j.reader.compile.parse.tagbuilder;
+package io.luchta.forma4j.reader.xml.parse.tagbuilder;
 
 import io.luchta.forma4j.context.syntax.SyntaxErrors;
 import io.luchta.forma4j.reader.model.tag.Tag;
