+++ src/main/java/io/luchta/forma4j/reader/xml/parse/tagbuilder/VForBuilder.java
@@ -1,9 +1,10 @@
-package io.luchta.forma4j.reader.compile.parse.tagbuilder;
+package io.luchta.forma4j.reader.xml.parse.tagbuilder;
 
 import io.luchta.forma4j.context.syntax.SyntaxError;
 import io.luchta.forma4j.context.syntax.SyntaxErrors;
-import io.luchta.forma4j.reader.model.tag.HForTag;
 import io.luchta.forma4j.reader.model.tag.Tag;
+import io.luchta.forma4j.reader.model.tag.VForTag;
+import io.luchta.forma4j.reader.model.tag.property.BreakCondition;
 import io.luchta.forma4j.reader.model.tag.property.Index;
 import io.luchta.forma4j.reader.model.tag.property.Name;
 import org.w3c.dom.NamedNodeMap;
@@ -12,13 +13,23 @@ import org.w3c.dom.Node;
 import java.util.ArrayList;
 import java.util.List;
 
-public class HForBuilder implements TagBuilder {
-
+/**
+ * {@code VForBuilder} はv-forタグの内容を読み取ってオブジェクト化するクラスです
+ *
+ * @since 0.1.0
+ */
+public class VForBuilder implements TagBuilder {
+    /**
+     * v-forタグのオブジェクト化を行います
+     * @param nodeMap
+     * @param syntaxErrors
+     * @return オブジェクト化されたv-forタグ。{@link io.luchta.forma4j.reader.model.tag.VForTag} に変換されます。
+     */
     @Override
     public Tag build(NamedNodeMap nodeMap, SyntaxErrors syntaxErrors) {
-
         List<String> errorMessages = new ArrayList<>();
 
+        // fromプロパティの読み込み
         Node fromNode = nodeMap.getNamedItem("from");
         boolean fromIsUndefined = false;
         if (fromNode == null) {
@@ -29,38 +40,37 @@ public class HForBuilder implements TagBuilder {
             errorMessages.add("from には0以上の整数を指定してください");
         }
 
+        // toプロパティの読み込み
         Node toNode = nodeMap.getNamedItem("to");
         boolean toIsUndefined = false;
         if (toNode == null) {
             toIsUndefined = true;
         }
-        Integer toNumber = convertNodeValueToInteger(toNode);
+        Integer toNumber = convertNodeValueToInteger(nodeMap.getNamedItem("to"));
         if (toNumber == null) {
             errorMessages.add("to には0以上の整数を指定してください");
         }
 
-        Node rowNode = nodeMap.getNamedItem("row");
-        boolean rowIsUndefined = false;
-        if (rowNode == null) {
-            rowIsUndefined = true;
-        }
-        Integer rowNumber = convertNodeValueToInteger(rowNode);
-        if (rowNumber == null) {
-            errorMessages.add("row には0以上の整数を指定してください");
-        }
-
+        // nameプロパティの読み込み
         String name = convertNodeValueToString(nodeMap.getNamedItem("name"));
         if (name == null) {
             errorMessages.add("name は必須入力です");
         }
 
+        // breakプロパティの読み込み
+        String breakCondition = convertNodeValueToString(nodeMap.getNamedItem("break"));
+
         addSyntaxErrors(errorMessages, syntaxErrors);
 
-        return new HForTag(new Index(fromNumber), fromIsUndefined, new Index(toNumber), toIsUndefined, new Index(rowNumber), rowIsUndefined, new Name(name));
+        return new VForTag(new Index(fromNumber), fromIsUndefined, new Index(toNumber), toIsUndefined, new Name(name), new BreakCondition(breakCondition));
     }
 
+    /**
+     * XMLから取得した値を数値に変換します
+     * @param node
+     * @return 取得値を数値に変換した値
+     */
     private Integer convertNodeValueToInteger(Node node) {
-
         Integer value = null;
         if (node == null) {
             value = 0;
@@ -79,8 +89,12 @@ public class HForBuilder implements TagBuilder {
         return value;
     }
 
+    /**
+     * XMLから取得した値を文字列に変換します
+     * @param node
+     * @return 取得値を文字列に変換した値
+     */
     private String convertNodeValueToString(Node node) {
-
         String value = "";
         if (node != null) {
             value = node.getNodeValue();
@@ -88,8 +102,12 @@ public class HForBuilder implements TagBuilder {
         return value;
     }
 
+    /**
+     * {@link io.luchta.forma4j.context.syntax.SyntaxErrors} へエラーを追加します
+     * @param errorMessages
+     * @param syntaxErrors
+     */
     private void addSyntaxErrors(List<String> errorMessages, SyntaxErrors syntaxErrors) {
-
         for (String message : errorMessages) {
             SyntaxError syntaxError = new SyntaxError(message, new UnsupportedOperationException());
             syntaxErrors.add(syntaxError);
