+++ src/main/java/io/luchta/forma4j/reader/xml/relation/taganalyzer/CellTagAnalyzer.java
@@ -1,4 +1,4 @@
-package io.luchta.forma4j.reader.compile.relation.taganalyzer;
+package io.luchta.forma4j.reader.xml.relation.taganalyzer;
 
 import io.luchta.forma4j.context.syntax.SyntaxError;
 import io.luchta.forma4j.context.syntax.SyntaxErrors;
@@ -7,7 +7,7 @@ import io.luchta.forma4j.reader.model.tag.TagTree;
 import io.luchta.forma4j.reader.model.tag.TagTrees;
 import io.luchta.forma4j.reader.model.tag.Tags;
 
-public class HForTagAnalyzer implements TagAnalyzer{
+public class CellTagAnalyzer implements TagAnalyzer {
 
     @Override
     public void analyze(TagTree tagTree, SyntaxErrors syntaxErrors) {
@@ -22,13 +22,13 @@ public class HForTagAnalyzer implements TagAnalyzer{
         }
 
         if (!hasSheet) {
-            SyntaxError syntaxError = new SyntaxError("h-for タグは sheet タグの子孫要素となるようにしてください");
+            SyntaxError syntaxError = new SyntaxError("cell タグは sheet タグの子孫要素となるようにしてください");
             syntaxErrors.add(syntaxError);
         }
 
         TagTrees children = tagTree.getChildren();
-        if (children.size() == 0) {
-            SyntaxError syntaxError = new SyntaxError("h-for タグには子要素が必要です");
+        if (children.size() != 0) {
+            SyntaxError syntaxError = new SyntaxError("cell タグは子要素を持つことができません");
             syntaxErrors.add(syntaxError);
         }
     }
