+++ src/main/java/io/luchta/forma4j/reader/xml/relation/taganalyzer/ListTagAnalyzer.java
@@ -1,4 +1,4 @@
-package io.luchta.forma4j.reader.compile.relation.taganalyzer;
+package io.luchta.forma4j.reader.xml.relation.taganalyzer;
 
 import io.luchta.forma4j.context.syntax.SyntaxError;
 import io.luchta.forma4j.context.syntax.SyntaxErrors;
@@ -6,20 +6,18 @@ import io.luchta.forma4j.reader.model.tag.Tag;
 import io.luchta.forma4j.reader.model.tag.TagTree;
 import io.luchta.forma4j.reader.model.tag.TagTrees;
 
-public class SheetTagAnalyzer implements TagAnalyzer {
-
+public class ListTagAnalyzer implements TagAnalyzer {
     @Override
     public void analyze(TagTree tagTree, SyntaxErrors syntaxErrors) {
-
         Tag immediateDominator = tagTree.getImmediateDominator();
-        if (!immediateDominator.isFormaReader()) {
-            SyntaxError syntaxError = new SyntaxError("sheet タグの親要素は forma-reader としてください");
+        if (!immediateDominator.isSheet()) {
+            SyntaxError syntaxError = new SyntaxError("list タグの親要素は sheet タグとしてください");
             syntaxErrors.add(syntaxError);
         }
 
         TagTrees children = tagTree.getChildren();
-        if (children.size() == 0) {
-            SyntaxError syntaxError = new SyntaxError("sheet タグには子要素が必要です");
+        if (children.size() != 0) {
+            SyntaxError syntaxError = new SyntaxError("list タグは子要素を持つことができません");
             syntaxErrors.add(syntaxError);
         }
     }
